efifb: arm/arm64: validate fb BAR instead of claiming it
Claiming the BAR that matches the framebuffer base address in a PCI
header fixup quirk only works as expected when the PCI device is on
bus 0. If not, it will produce the following error:

pci 0000:01:01.0: can't claim BAR 0 [mem 0x10000000-0x10ffffff pref]:
                                                 no compatible bridge window
pci 0000:01:01.0: BAR 0: failed to claim resource for efifb!

Since claiming the entire path up to the root is non-trivial at this
point, and also undesirable given that we have no way of making sure
that the existing configuration is sound, let's not claim anything at
all, and simply record the resource so we can check that it did not
move when the EFI fb driver is probed.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: knguyen <knguyen@amcc.com>