BACKPORT: ACPI / scan: Apply default enumeration to devices with ACPI drivers
The current code in acpi_bus_attach() is inconsistent with respect
to device objects with ACPI drivers bound to them, as it allows
ACPI drivers to bind to device objects with existing "physical"
device companions, but it doesn't allow "physical" device objects
to be created for ACPI device objects with ACPI drivers bound to
them.  Thus, in some cases, the outcome depends on the ordering
of events which is confusing at best.

For this reason, modify acpi_bus_attach() to call
acpi_default_enumeration() for device objects with the
pnp.type.platform_id flag set regardless of whether or not
any ACPI drivers are bound to them.

This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git/commit/?h=next-20171018&id=f5beabfe61794d9a9d9549d387cda2bffd81e504

Without the patch, ACPI devices declared in DSDT would not be enumerated,
hence, both I2C MUX/LED on Merlin and SSIF of OPEN IPMI on Ospey do not work.

Change-Id: I790d367175d429ddbce205ba2b89447f19618968
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Reviewed-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Reviewed-by: Joey Lee <jlee@suse.com>
Signed-off-by: Tin Huynh <tnhuynh@apm.com>