BACKPORT: ata: fixes kernel crash while tracing ata_eh_link_autopsy event
When tracing ata link error event, the kernel crashes when the disk is
removed due to NULL pointer access by trace_ata_eh_link_autopsy API.
This occurs as the dev is NULL when the disk disappeared. This patch
fixes this crash by calling trace_ata_eh_link_autopsy only if "dev"
is not NULL.

Patch reference:
https://git.kernel.org/pub/scm/linux/kernel/git/tj/libata.git/commit/?h=for-4.15&id=f1601113ddc0339a745e702f4fb1ca37d4875e65

This patch is required to support user space RAS for X-Gene hardware v3
for PCIe and SATA protocols.

Change-Id: I63747a4017f95cfb2976253ef1f0422991073887
Signed-off-by: Rameshwar Prasad Sahu <rsahu@apm.com>